Change `PremiumLeft` type from `long` to `double` and update `FormatSeconds` method parameter type

+++ src/Domain/Account/Infos/AccountInfo.cs
@@ -3,7 +3,7 @@ public sealed class AccountInfo
 	public string Username { get; init; } = default!;
 	public string Email { get; init; } = default!;
 	public bool EmailVerified { get; init; }
-	public long PremiumLeft { get; init; }
+	public double PremiumLeft { get; init; }
 	public int Pts { get; init; }
 
 	public Uri UpgradeAccountUrl { get; init; } = default!;
